Step one and two
Change the index page so you can write in a term and click "Search" or
press enter and you will go to the results.php page. You should see your
term in the url/address bar. Hint: use a form.
Do the same thing for the text field on the results page.